requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
asgiref==3.4.1
astroid==2.8.4
async-generator==1.10
async-timeout==4.0.0
attrs==21.2.0
autopep8==1.6.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
cssselect==1.1.0
decorator==5.1.0
dicttoxml==1.7.4
Django==3.2.8
django-ipware==4.0.0
docutils==0.18
frozenlist==1.2.0
future==0.18.2
geocoder==1.38.1
geoip2==4.4.0
gunicorn==20.1.0
h11==0.12.0
idna==3.3
importlib-metadata==4.8.1
ip2geotools==0.1.6
IP2Location==8.6.2
isort==5.10.0
jeepney==0.7.1
keyring==23.2.1
lazy-object-proxy==1.6.0
lxml==4.6.4
maxminddb==2.2.0
mccabe==0.6.1
multidict==5.2.0
outcome==1.1.0
packaging==21.2
pip-review==1.1.0
pkginfo==1.7.1
platformdirs==2.4.0
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.10.0
pylint==2.11.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyquery==1.4.3
pytz==2021.3
ratelim==0.1.6
readme-renderer==30.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.1
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
twine==3.5.0
typed-ast==1.4.3
typing==3.7.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
webencodings==0.5.1
wrapt==1.13.3
wsproto==1.0.0
yarl==1.7.2
zipp==3.6.0
